Fixed more bugs, working time remaining.

+++ ASDocument.h
@@ -26,7 +26,7 @@
 	ASWindowController *windowController;
 	ASSFVData *_data;
 	NSDate *_date;
-	long _dataRead;
+	long long _dataRead;
 	float _percentCompleted;
 	float _filePercentCompleted;
 	BOOL _threadShouldExit;
